Bump the disk_allocation_ratio for the compute nodes
We're not changing the default values for cpu and ram but we are
making them explicit.
We're bumping the disk_allocation_ratio a bit because it's mostly
our scheduling bottleneck and we can spare to overcommit due to
qcow2 thin provisioning /and/ planning on tenants not really using
all the space they're allocated.