Manage kernel modules with puppet-kmod instead of "manually"
puppet-kmod will take care of reloading the module when the
initial configuration is done.
For example, the previous way would configure kvm_intel nested to
be enabled but would not reload the module, thus the nested flag
was actually false and the vmx flag was not exposed to the virtual
machines.